Remove InterpolatedProperty from Technique.
InterpolatedProperty is only used in scope of interpolation
operators so can be removed from typings and "general" codebase.
