+++ @here/harp-datasource-protocol/lib/InterpolatedPropertyDefs.ts
@@ -5,7 +5,6 @@
  */
 
 import { JsonExpr } from "./Expr";
-import { StringEncodedNumeralType } from "./StringEncodedNumeral";
 
 /**
  * Interpolation mode used when computing a [[InterpolatedProperty]] value for a given zoom level.
@@ -36,41 +35,26 @@ export interface InterpolatedPropertyDefinition<T> {
 }
 
 /**
- * Property which value is interpolated across different zoom levels.
+ * Checks if a property is interpolated.
+ * @param p property to be checked
  */
-export interface InterpolatedProperty {
-    /**
-     * Interpolation mode that should be used for this property.
-     */
-    interpolationMode: InterpolationMode;
-
-    /**
-     * Zoom level keys array.
-     */
-    zoomLevels: Float32Array;
-
-    /**
-     * Property values array.
-     */
-    values: ArrayLike<any>;
-
-    /**
-     * Exponent used in interpolation. Only valid with `Exponential` [[InterpolationMode]].
-     */
-    exponent?: number;
-
-    /**
-     * @hidden
-     * [[StringEncodedNumeral]] type needed to interpret interpolated values back to numbers.
-     */
-    _stringEncodedNumeralType?: StringEncodedNumeralType;
-
-    /**
-     * @hidden
-     * Array of `0` and `1`mask values used to modify the interpolation behaviour of some
-     * [[StringEncodedNumeral]]s.
-     */
-    _stringEncodedNumeralDynamicMask?: Float32Array;
+export function isInterpolatedPropertyDefinition<T>(
+    p: any
+): p is InterpolatedPropertyDefinition<T> {
+    if (
+        p &&
+        p.interpolationMode === undefined &&
+        Array.isArray(p.values) &&
+        p.values.length > 0 &&
+        p.values[0] !== undefined &&
+        Array.isArray(p.zoomLevels) &&
+        p.zoomLevels.length > 0 &&
+        p.zoomLevels[0] !== undefined &&
+        p.values.length === p.zoomLevels.length
+    ) {
+        return true;
+    }
+    return false;
 }
 
 /**
