+++ @here/harp-datasource-protocol/lib/TechniqueAttr.ts
@@ -6,8 +6,6 @@
 
 import { LoggerManager } from "@here/harp-utils";
 import { Env, Expr, ExprScope, MapEnv, Value } from "./Expr";
-import { getPropertyValue, isInterpolatedProperty } from "./InterpolatedProperty";
-import { InterpolatedProperty } from "./InterpolatedPropertyDefs";
 
 const logger = LoggerManager.instance.create("TechniqueAttr");
 
@@ -46,7 +44,7 @@ export interface AttrEvaluationContext {
  */
 export function evaluateTechniqueAttr<T = Value>(
     context: Env | AttrEvaluationContext,
-    attrValue: T | Expr | InterpolatedProperty | undefined
+    attrValue: T | Expr | undefined
 ): T | undefined;
 
 /**
@@ -56,13 +54,13 @@ export function evaluateTechniqueAttr<T = Value>(
  */
 export function evaluateTechniqueAttr<T = Value>(
     context: Env | AttrEvaluationContext,
-    attrValue: T | Expr | InterpolatedProperty | undefined,
+    attrValue: T | Expr | undefined,
     defaultValue: T
 ): T;
 
 export function evaluateTechniqueAttr<T = Value>(
     context: Env | AttrEvaluationContext,
-    attrValue: T | Expr | InterpolatedProperty | undefined,
+    attrValue: T | Expr | undefined,
     defaultValue?: T
 ): T | undefined {
     const env = context instanceof Env ? context : context.env;
@@ -79,8 +77,6 @@ export function evaluateTechniqueAttr<T = Value>(
             logger.error(`failed to evaluate expression '${JSON.stringify(attrValue)}': ${error}`);
             evaluated = undefined;
         }
-    } else if (isInterpolatedProperty(attrValue)) {
-        evaluated = getPropertyValue(attrValue, context instanceof Env ? context : context.env);
     } else {
         evaluated = (attrValue as unknown) as Value;
     }
